Add --sleep option to rate limit CardDAV requests
As the script is too fast for the rate limiting of my CardDAV server (Monica), I've added an option to slow down in between VCard calls. The command line option takes a float (so 0.5 works as well) and specifies the number of seconds to sleep. Defaults to 0 to retain current behaviour.